Update blog_post.md
In the beginning I mentioned that coroutines allow you to write asynchronous code the same way as you would with synchronous code: "with" is not needed -> 
In the beginning I mentioned that coroutines allow you to write asynchronous code the same way as you would synchronous code
maybe hard to understand. Hence my proposal:
In the beginning I mentioned that coroutines allow you to write asynchronous code that is as easy to read as synchronous code.